backend/routes/auth_routes.py
"""
Authentication routes
"""
from flask import Blueprint, request, jsonify
from werkzeug.security import generate_password_hash, check_password_hash
import jwt
import os
from datetime import datetime, timedelta

bp = Blueprint('auth', __name__, url_prefix='/api/auth')

@bp.route('/register', methods=['POST'])
def register():
    """Register a new user"""
    data = request.get_json()
    
    if not data or not data.get('email') or not data.get('password'):
        return jsonify({'error': 'Missing email or password'}), 400
    
    # TODO: Implement user registration with database
    return jsonify({'message': 'User registered successfully'}), 201

@bp.route('/login', methods=['POST'])
def login():
    """Login user"""
    data = request.get_json()
    
    if not data or not data.get('email') or not data.get('password'):
        return jsonify({'error': 'Missing email or password'}), 400
    
    # TODO: Implement user login with token generation
    return jsonify({'token': 'jwt-token-here'}), 200

@bp.route('/logout', methods=['POST'])
def logout():
    """Logout user"""
    # TODO: Implement logout functionality
    return jsonify({'message': 'Logged out successfully'}), 200

@bp.route('/verify', methods=['POST'])
def verify_token():
    """Verify JWT token"""
    token = request.headers.get('Authorization', '').replace('Bearer ', '')
    
    if not token:
        return jsonify({'error': 'Missing token'}), 401
    
    # TODO: Implement token verification
    return jsonify({'valid': True}), 200
