backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
PyJWT==2.10.1
requests==2.31.0
openai==1.3.0
zxcvbn==4.5.0
bcrypt==4.1.1
Werkzeug==3.0.1
pymongo==4.6.0
